+++ android/src/main/java/com/viztushar/flutter_webview_mod/FlutterWebView.java
@@ -285,6 +285,8 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
                     });
           }
           break;
+          case "gestureNavigationEnabled":
+          break;
         case "userAgent":
           updateUserAgent((String) settings.get(key));
           break;